requirements.txt
flask
gunicorn
tensorflow
numpy
pillow
opencv-python-headless
matplotlib
gdown